Refactor `ATNSimulator` into a parser trait, remove static ambiguity detection and add tests
