+++ packages/chevrotain/src/parse/parser/traits/gast_recorder.ts
@@ -416,13 +416,6 @@ function recordOrProd(mainProdArg: any, occurrence: number): any {
       predicate: currAlt.GATE
     })
     newOrProd.definition.push(currAltFlat)
-    if (has(currAlt, "IGNORE_AMBIGUITIES")) {
-      currAltFlat.ignoreAmbiguities = currAlt.IGNORE_AMBIGUITIES as boolean // assumes end user provides the correct config value/type
-    }
-    // **implicit** ignoreAmbiguities due to usage of gate
-    else if (has(currAlt, "GATE")) {
-      currAltFlat.ignoreAmbiguities = true
-    }
     this.recordingProdStack.push(currAltFlat)
     currAlt.ALT.call(this)
     this.recordingProdStack.pop()
