+++ packages/chevrotain/test/parse/grammar/checks_spec.ts
@@ -1208,150 +1208,6 @@ describe("The empty alternative detection full flow", () => {
     expect(() => new EmptyAltIndirectAmbiguityParser()).to.throw("2")
   })
 
-  it("will detect alternative ambiguity with identical lookaheads", () => {
-    class AltAmbiguityParserImplicitOccurence extends EmbeddedActionsParser {
-      constructor(input: IToken[] = []) {
-        super([PlusTok, StarTok])
-        this.performSelfAnalysis()
-        this.input = input
-      }
-
-      public noneLastEmpty = this.RULE("noneLastEmpty", () => {
-        this.OR([
-          {
-            ALT: () => {
-              this.CONSUME1(PlusTok)
-              this.CONSUME1(StarTok)
-            }
-          },
-          {
-            ALT: () => {
-              this.CONSUME2(PlusTok)
-              this.CONSUME2(StarTok)
-            }
-          }
-        ])
-      })
-    }
-    expect(() => new AltAmbiguityParserImplicitOccurence()).to.throw(
-      "Ambiguous Alternatives Detected"
-    )
-    expect(() => new AltAmbiguityParserImplicitOccurence()).to.throw("1")
-    expect(() => new AltAmbiguityParserImplicitOccurence()).to.throw("2")
-    expect(() => new AltAmbiguityParserImplicitOccurence()).to.throw(
-      "<PlusTok, StarTok> may appears as a prefix path"
-    )
-  })
-
-  it("will detect alternative ambiguity with identical lookahead - custom maxLookAhead", () => {
-    class AltAmbiguityParserImplicitOccurrence extends EmbeddedActionsParser {
-      constructor(input: IToken[] = []) {
-        super([PlusTok, StarTok])
-        this.performSelfAnalysis()
-        this.input = input
-      }
-
-      public noneLastEmpty = this.RULE(
-        "AmbiguityDueToExplicitLowLookahead",
-        () => {
-          this.OR({
-            MAX_LOOKAHEAD: 1,
-            DEF: [
-              {
-                ALT: () => {
-                  this.CONSUME1(PlusTok)
-                  this.CONSUME1(PlusTok)
-                }
-              },
-              {
-                ALT: () => {
-                  this.CONSUME2(PlusTok)
-                  this.CONSUME2(StarTok)
-                }
-              }
-            ]
-          })
-        }
-      )
-    }
-    expect(() => new AltAmbiguityParserImplicitOccurrence()).to.throw(
-      "Ambiguous Alternatives Detected"
-    )
-    expect(() => new AltAmbiguityParserImplicitOccurrence()).to.throw("1")
-    expect(() => new AltAmbiguityParserImplicitOccurrence()).to.throw("2")
-    expect(() => new AltAmbiguityParserImplicitOccurrence()).to.throw(
-      "<PlusTok> may appears as a prefix path"
-    )
-  })
-
-  context("IGNORE_AMBIGUITIES flag", () => {
-    it("will ignore specific alternative ambiguity", () => {
-      class IgnoreAlternativeAmbiguitiesFlagParser extends EmbeddedActionsParser {
-        constructor(input: IToken[] = []) {
-          super([PlusTok, StarTok])
-          this.performSelfAnalysis()
-          this.input = input
-        }
-
-        public noneLastEmpty = this.RULE("noneLastEmpty", () => {
-          this.OR([
-            {
-              IGNORE_AMBIGUITIES: true,
-              ALT: () => {
-                this.CONSUME1(PlusTok)
-                this.CONSUME1(StarTok)
-              }
-            },
-            {
-              ALT: () => {
-                this.CONSUME2(PlusTok)
-                this.CONSUME2(StarTok)
-              }
-            }
-          ])
-        })
-      }
-      expect(() => new IgnoreAlternativeAmbiguitiesFlagParser()).to.not.throw()
-    })
-
-    it("will ignore all alternation ambiguities", () => {
-      class IgnoreAlternationAmbiguitiesFlagParser extends EmbeddedActionsParser {
-        constructor(input: IToken[] = []) {
-          super([PlusTok, StarTok])
-          this.performSelfAnalysis()
-          this.input = input
-        }
-
-        public noneLastEmpty = this.RULE("noneLastEmpty", () => {
-          this.OR({
-            IGNORE_AMBIGUITIES: true,
-            DEF: [
-              {
-                ALT: () => {
-                  this.CONSUME1(PlusTok)
-                  this.CONSUME1(StarTok)
-                }
-              },
-              {
-                ALT: () => {
-                  this.CONSUME2(PlusTok)
-                  this.CONSUME2(StarTok)
-                }
-              },
-              {
-                ALT: () => {
-                  this.CONSUME3(PlusTok)
-                  this.CONSUME3(StarTok)
-                }
-              }
-            ]
-          })
-        })
-      }
-      expect(() => new IgnoreAlternationAmbiguitiesFlagParser()).to.not.throw()
-    })
-  })
-
   it("will throw an error when an empty alternative is not the last alternative #2", () => {
     class EmptyAltAmbiguityParser2 extends EmbeddedActionsParser {
       constructor(input: IToken[] = []) {
@@ -1387,186 +1243,6 @@ describe("The empty alternative detection full flow", () => {
       "Only the last alternative may be an empty alternative."
     )
     expect(() => new EmptyAltAmbiguityParser2()).to.not.throw("undefined")
-  })
-})
-
-describe("The prefix ambiguity detection full flow", () => {
-  it("will throw an error when an a common prefix ambiguity is detected - categories", () => {
-    const A = createToken({ name: "A" })
-    const B = createToken({ name: "B", categories: A })
-    const C = createToken({ name: "C" })
-
-    class PrefixAltAmbiguity extends EmbeddedActionsParser {
-      constructor(input: IToken[] = []) {
-        super([A, B, C])
-        this.performSelfAnalysis()
-        this.input = input
-      }
-
-      public prefixAltAmbiguity = this.RULE("prefixAltAmbiguity", () => {
-        this.OR3([
-          {
-            ALT: () => {
-              this.CONSUME1(B)
-              this.CONSUME1(A)
-            }
-          },
-          {
-            ALT: () => {
-              this.CONSUME2(A)
-              this.CONSUME3(A)
-              this.CONSUME1(C)
-            }
-          }
-        ])
-      })
-    }
-    expect(() => new PrefixAltAmbiguity()).to.throw("OR3")
-    expect(() => new PrefixAltAmbiguity()).to.throw("Ambiguous alternatives")
-    expect(() => new PrefixAltAmbiguity()).to.throw(
-      "due to common lookahead prefix"
-    )
-    expect(() => new PrefixAltAmbiguity()).to.throw("<B, A>")
-    expect(() => new PrefixAltAmbiguity()).to.throw(
-      "https://chevrotain.io/docs/guide/resolving_grammar_errors.html#COMMON_PREFIX"
-    )
-  })
-
-  it("will throw an error when an an alts ambiguity is detected", () => {
-    const OneTok = createToken({ name: "OneTok" })
-    const TwoTok = createToken({ name: "TwoTok" })
-    const Comma = createToken({ name: "Comma" })
-
-    const ALL_TOKENS = [OneTok, TwoTok, Comma]
-
-    class AlternativesAmbiguityParser extends EmbeddedActionsParser {
-      constructor() {
-        super(ALL_TOKENS, {
-          maxLookahead: 4
-        })
-        this.performSelfAnalysis()
-      }
-
-      public main = this.RULE("main", () => {
-        this.OR([
-          { ALT: () => this.SUBRULE(this.alt1) },
-          { ALT: () => this.SUBRULE(this.alt2) }
-        ])
-      })
-
-      public alt1 = this.RULE("alt1", () => {
-        this.MANY(() => {
-          this.CONSUME(Comma)
-        })
-        this.CONSUME(OneTok)
-      })
-
-      public alt2 = this.RULE("alt2", () => {
-        this.MANY(() => {
-          this.CONSUME(Comma)
-        })
-        this.CONSUME(TwoTok)
-      })
-    }
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "Ambiguous Alternatives Detected: <1 ,2>"
-    )
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "in <OR> inside <main> Rule"
-    )
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "Comma, Comma, Comma, Comma"
-    )
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "https://chevrotain.io/docs/guide/resolving_grammar_errors.html#AMBIGUOUS_ALTERNATIVES"
-    )
-  })
-
-  it("will throw an error when an an alts ambiguity is detected - Categories", () => {
-    const A = createToken({ name: "A" })
-    const B = createToken({ name: "B" })
-    const C = createToken({ name: "C" })
-    const D = createToken({ name: "D", categories: C })
-
-    const ALL_TOKENS = [A, B, C, D]
-
-    class AlternativesAmbiguityParser extends EmbeddedActionsParser {
-      constructor() {
-        super(ALL_TOKENS, {
-          maxLookahead: 4
-        })
-        this.performSelfAnalysis()
-      }
-
-      public main = this.RULE("main", () => {
-        this.OR([
-          { ALT: () => this.SUBRULE(this.alt1) },
-          { ALT: () => this.SUBRULE(this.alt2) }
-        ])
-      })
-
-      public alt1 = this.RULE("alt1", () => {
-        this.MANY(() => {
-          this.CONSUME(D)
-        })
-        this.CONSUME(A)
-      })
-
-      public alt2 = this.RULE("alt2", () => {
-        this.MANY(() => {
-          this.CONSUME(C)
-        })
-        this.CONSUME(B)
-      })
-    }
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "Ambiguous Alternatives Detected: <1 ,2>"
-    )
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "in <OR> inside <main> Rule"
-    )
-    expect(() => new AlternativesAmbiguityParser()).to.throw("D, D, D, D")
-    expect(() => new AlternativesAmbiguityParser()).to.throw(
-      "https://chevrotain.io/docs/guide/resolving_grammar_errors.html#AMBIGUOUS_ALTERNATIVES"
-    )
-  })
-
-  // TODO: detect these ambiguity with categories
-  it("will throw an error when an a common prefix ambiguity is detected - implicit occurrence idx", () => {
-    class PrefixAltAmbiguity2 extends EmbeddedActionsParser {
-      constructor(input: IToken[] = []) {
-        super([PlusTok, MinusTok, StarTok])
-        this.performSelfAnalysis()
-        this.input = input
-      }
-
-      public prefixAltAmbiguity = this.RULE("prefixAltAmbiguity", () => {
-        this.OR([
-          {
-            ALT: () => {
-              this.CONSUME1(PlusTok)
-              this.CONSUME1(MinusTok)
-            }
-          },
-          {
-            ALT: () => {
-              this.CONSUME2(PlusTok)
-              this.CONSUME2(MinusTok)
-              this.CONSUME1(StarTok)
-            }
-          }
-        ])
-      })
-    }
-    expect(() => new PrefixAltAmbiguity2()).to.throw("OR")
-    expect(() => new PrefixAltAmbiguity2()).to.throw("Ambiguous alternatives")
-    expect(() => new PrefixAltAmbiguity2()).to.throw(
-      "due to common lookahead prefix"
-    )
-    expect(() => new PrefixAltAmbiguity2()).to.throw("<PlusTok, MinusTok>")
-    expect(() => new PrefixAltAmbiguity2()).to.throw(
-      "https://chevrotain.io/docs/guide/resolving_grammar_errors.html#COMMON_PREFIX"
-    )
   })
 })
