+++ packages/chevrotain/test/parse/recognizer_lookahead_spec.ts
@@ -931,101 +931,6 @@ describe("lookahead Regular Tokens Mode", () => {
     })
   })
 
-  describe("OR production ambiguity detection when using implicit lookahead calculation", () => {
-    it("will throw an error when two alternatives have the same single token (lookahead 1) prefix", () => {
-      class OrAmbiguityLookAheadParser extends EmbeddedActionsParser {
-        constructor(input: IToken[] = []) {
-          super(ALL_TOKENS, {})
-
-          this.performSelfAnalysis()
-          this.input = input
-        }
-
-        public ambiguityRule = this.RULE(
-          "ambiguityRule",
-          this.parseAmbiguityRule
-        )
-
-        private parseAmbiguityRule(): void {
-          this.OR1([
-            {
-              ALT: () => {
-                this.CONSUME1(OneTok)
-              }
-            },
-            // <-- this alternative starts with the same token as the previous one, ambiguity!
-            {
-              ALT: () => {
-                this.CONSUME2(OneTok)
-              }
-            },
-            {
-              ALT: () => {
-                this.CONSUME2(TwoTok)
-              }
-            },
-            {
-              ALT: () => {
-                this.CONSUME2(ThreeTok)
-              }
-            }
-          ])
-        }
-      }
-
-      expect(() => new OrAmbiguityLookAheadParser()).to.throw(
-        "Ambiguous Alternatives Detected:"
-      )
-      expect(() => new OrAmbiguityLookAheadParser()).to.throw("OneTok")
-    })
-
-    it("will throw an error when two alternatives have the same multi token (lookahead > 1) prefix", () => {
-      class OrAmbiguityMultiTokenLookAheadParser extends EmbeddedActionsParser {
-        constructor(input: IToken[] = []) {
-          super(ALL_TOKENS, {})
-
-          this.performSelfAnalysis()
-          this.input = input
-        }
-
-        public ambiguityRule = this.RULE(
-          "ambiguityRule",
-          this.parseAmbiguityRule
-        )
-
-        private parseAmbiguityRule(): void {
-          this.OR1([
-            {
-              ALT: () => {
-                this.CONSUME1(OneTok)
-              }
-            },
-            {
-              ALT: () => {
-                this.CONSUME1(TwoTok)
-                this.CONSUME1(ThreeTok)
-                this.CONSUME1(FourTok)
-              }
-            },
-            {
-              ALT: () => {
-                this.CONSUME2(TwoTok)
-                this.CONSUME2(ThreeTok)
-                this.CONSUME2(FourTok)
-              }
-            }
-          ])
-        }
-      }
-      expect(() => new OrAmbiguityMultiTokenLookAheadParser()).to.throw(
-        "Ambiguous Alternatives Detected:"
-      )
-      expect(() => new OrAmbiguityMultiTokenLookAheadParser()).to.throw(
-        "TwoTok, ThreeTok, FourTok"
-      )
-    })
-  })
-
   describe("The implicit lookahead calculation functionality of the Recognizer For OR (with IGNORE_AMBIGUITIES)", () => {
     class OrImplicitLookAheadParserIgnoreAmbiguities extends EmbeddedActionsParser {
       constructor(input: IToken[] = []) {
